feat: change parse for vbo ebo

+++ src/modules/include/matrix.h
@@ -19,7 +19,7 @@
 #define FAILURE 0
 
 typedef struct {
-  double *matrix;
+  float *matrix;
   int rows;
   int cols;
 } matrix_t;
